split prompts file
